Feature/create plant

+++ pages/plants/[id].js
@@ -49,20 +49,22 @@ export default function PlantDetailPage() {
     }
   }
 
-  async function handleDelete() {
-    const response = await fetch(`/api/plants/${id}`, {
-      method: "DELETE",
-    });
-    if (response.ok) router.push("/");
+  async function handleDelete(id) {
+    const isConfirmed = window.confirm(
+    `Are you sure you want to delete "${plant.name}"?`
+  );
+  if (!isConfirmed) {
+    return;
   }
-
+  await fetch(`/api/plants/${id}`, {
+    method: "DELETE",
+  });
+  router.push("/");
+}
   return (
     <>
-      <PlantDetails plant={plant} onEdit={handleEdit} />
-
-      <button type="button" onClick={handleDelete}>
-        Delete
-      </button>
+      <PlantDetails plant={plant} onEdit={handleEdit} onDelete={handleDelete}/>
+      
     </>
   );
 }
